requirements.txt
crewai==0.28.8
crewai-tools==0.1.6
gradio==4.31.5
openai==1.30.0
python-dotenv==1.0.1
duckduckgo-search==5.3.0